refactor: review top bar components state and styles
