+++ src/app/@topbar/_components/skip-button.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Button } from '@/app/pwa/_components/ui/button'
+import { Button } from '@/app/_components/ui/button'
 import { useRouter } from 'next/navigation'
 
 export default function SkipButton() {
