+++ src/app/@topbar/_components/user-dropdown.tsx
@@ -11,7 +11,7 @@ import {
     DropdownMenuContent,
     DropdownMenuPortal,
     DropdownMenuTrigger,
-} from '@/app/pwa/_components/ui/dropdown-menu'
+} from '@/app/_components/ui/dropdown-menu'
 import { EllipsisIcon } from 'lucide-react'
 import { useEffect, useState } from 'react'
 import UserDropdownList from './user-dropdown.list'
